FEAT: pipeline de testes e ajustes em testes
