+++ tests/conftest.py
@@ -1,5 +1,4 @@
 import pytest
-from flask import json
 
 from src.app import create_app
 from src.app.routes import routes
@@ -17,7 +16,7 @@ def app():
     app_on = create_app('testing')
     routes(app_on)
     return app_on
-    
+
 
 @pytest.fixture
 def logged_in_client(client):
